# docs/swagger.yaml
basePath: /
definitions:
...
  models.SMSConfirmation:
    properties:
      confirmed:
        type: boolean
      confirmed_at:
        type: string
      document_id:
        type: integer
      id:
        type: integer
      phone:
        type: string
      sent_at:
        type: string
      sms_code:
        type: string
    type: object
...
host: localhost:4000
info:
  contact:
    email: support@company.com
    name: Тур Компани
    url: http://contact.company.com
  description: Aпишка для управления туристической компанией (документация Swagger).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://your-terms.com/terms/
  title: Tour Company API
  version: "1.0"
paths:
...
  /sms/{document_id}:
    delete:
      description: Удаляет все SMS-подтверждения по документу
      parameters:
      - description: ID документа
        in: path
        name: document_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить SMS-подтверждения
      tags:
      - SMS
    get:
      description: Возвращает последнее SMS по документу
      parameters:
      - description: ID документа
        in: path
        name: document_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SMSConfirmation'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить последнее SMS
      tags:
      - SMS
  /sms/confirm:
    post:
      consumes:
      - application/json
      description: Подтверждает введённый код по ID документа
      parameters:
      - description: ID документа и код подтверждения
        in: body
        name: input
        required: true
        schema:
          properties:
            code:
              type: string
            document_id:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Подтвердить SMS-код
      tags:
      - SMS
  /sms/resend:
    get:
      description: Повторно отправляет SMS по ID документа
      parameters:
      - description: ID документа
        in: query
        name: document_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Повторная отправка SMS
      tags:
      - SMS
  /sms/send:
    post:
      consumes:
      - application/json
      description: Отправляет SMS с кодом подтверждения на указанный номер
      parameters:
      - description: Данные для отправки SMS
        in: body
        name: input
        required: true
        schema:
          properties:
            document_id:
              type: integer
            phone:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Отправить SMS
      tags:
      - SMS
...
swagger: "2.0"